pyproject.toml
[build-system]
# Specify the build system and dependencies required for building the project
requires = ["hatchling", "versioningit"]
build-backend = "hatchling.build"

[project]
# Define the project metadata
name = "connector"
description = "Kafka Connector Management tool"
authors = []
requires-python = ">= 3.10"
readme = "README.md"
license = {file = "LICENSE"}
include = ["pyproject.toml", "README.md", "LICENSE"]

# List the project dependencies
dependencies = [
    "python-dotenv>=1.0.1",
]

# Specify dynamic versioning
dynamic = ["version"]

[project.optional-dependencies]
# Define optional dependencies for the project
docs = [
]

[tool.hatch.version]
# Specify the versioning source tool
source = "versioningit"

[tool.versioningit.next-version]
# Define the versioning method
method = "smallest"


[tool.versioningit.format]
# Define the format for the versioningit tool
distance = "{next_version}.dev{distance}"
dirty = "{next_version}.dev{distance}"
distance-dirty = "{next_version}.dev{distance}"

[tool.hatch.metadata]
# Specify dynamic metadata fields for Hatch
dynamic = ["version"]

[project.scripts]
# Define the entry point for the CLI
connector_cli = "connector.cli:main"